Optimize flake areas
So far, we've been drawing all flakes in canvas of the size of the
original randomly drawn flake.

All following flakes of the size-canvas-array are smaller than that.

Now we just use the smaller size of the projected flake image plus
2 times the current blur value as size.

This has the added advantage that we keep the calculated flake sizes
as canvas sizes. We will need that for our next step.