Handle window resizing
Add a debounced window resize handler that reinitializes the canvas to the
new target size.

We just let the flakes do their thing. This leads to irregularities in the
random distribution. But this seems preferable to totally reinitializing
the scene. The wonky distribution also corrects itself over time.

This restructures the width/height handling a bit. Instead of keeping
the screen width in every flake we just use a closure which I should have
done in the first place. This spares us from having to update the screen
limits in every flake.